Make TIOCGDEV compatible again.
suse-commit: 30edde304eb8287a41a11dc3e455c531d6219ce9

+++ drivers/char/tty_io.c
@@ -1892,8 +1892,8 @@ int tty_ioctl(struct inode * inode, struct file * file,
 		 */
 		case TIOCGDEV:
 		{
-			unsigned long ret = new_encode_dev(tty_devnum(real_tty));
-			return put_user(ret, (unsigned long*) arg);
+			unsigned int ret = new_encode_dev(tty_devnum(real_tty));
+			return put_user(ret, (unsigned int*) arg);
 		}
 
 	}
